requirements.txt
caldav==2.0.1
faster_whisper==1.1.1
langchain==0.3.26
langchain_core==0.3.69
langchain_huggingface==0.3.0
langchain_ollama==0.3.5
langchain_qdrant==0.2.0
langgraph==0.5.3
numpy==2.3.1
pandas==2.3.1
prompt_toolkit==3.0.48
py3_validate_email==1.0.5.post2
python-dotenv==1.1.1
qdrant_client==1.14.3
scipy==1.16.0
sounddevice==0.5.2
streamlit==1.46.0
tqdm==4.67.0
TTS
sounddevice==0.5.2
